Move and cleanup ActivityDetails
