+++ ui/src/AtlasUI/components/activity/ActivityDetails/ActivityDetails.module.css
@@ -1,0 +1,3 @@
+.activityDetails {
+  padding: 1em;
+}
