+++ ui/src/AtlasUI/components/activity/ActivityDetails/ContainerCollection.js
