+++ ui/src/AtlasUI/components/activity/ActivityDetails/index.js
