+++ ui/src/AtlasUI/components/index.js
@@ -1,2 +1,3 @@
 export { ActivitySearch } from './activity/ActivitySearch';
 export { ActivityHeader } from './activity/ActivityHeader';
+export { ActivityDetails } from './activity/ActivityDetails';
