--- ui/src/components/activity/ActivityDetails/ActivityDetails.module.css
@@ -1,28 +1,0 @@
-.experimentDetails {
-  padding: 1em;
-}
-
-.loader {
-  margin-top: 4em;
-}
-
-.newActivityContainer {
-  display: flex;
-  justify-content: center;
-}
-
-.newActivity {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-}
-
-.newActivity :global(h3) {
-  margin-bottom: 0 !important;
-}
-
-.plateSizeFormContainer {
-  margin-top: 3em;
-  margin-bottom: 3em;
-}
