--- ui/src/components/activity/ContainerCollection/index.js
@@ -1,1 +1,0 @@
-export { ContainerCollection } from './ContainerCollection';
