+++ ui/src/components/index.js
@@ -1,5 +1,4 @@
 export { Header } from './app/Header';
-export { ActivityDetails } from './activity/ActivityDetails';
 export { Print } from './print';
 export { ComponentList } from './editor/ComponentList';
 export { Tools } from './editor/Tools';
