+++ ui/src/implementations/kapture-backend/components/ActivityDetailsConnected.js
@@ -1,26 +1,22 @@
 import { connect } from 'react-redux';
 
-import { ActivityDetails } from '../../../components';
+import { ActivityDetails } from 'AtlasUI/components';
 import { selectors } from '../store';
 import { actions } from '../actions';
 
+const { selectActivityContainerCollections } = selectors;
 const { setContainerCollectionsStale } = actions.activity;
-
-const { selectActivityName, selectActivityContainerCollections } = selectors;
 
 const mapState = (state, props) => {
   return {
-    activityName: selectActivityName(state),
     containerCollections: selectActivityContainerCollections(state),
   };
 };
 
-const onUnmount = () => {
-  return setContainerCollectionsStale({ stale: true });
-};
-
 const mapDispatch = {
-  onUnmount,
+  onUnmount: () => {
+    return setContainerCollectionsStale({ stale: true });
+  },
 };
 
 const connected = connect(mapState, mapDispatch)(ActivityDetails);
